Move set up and tear down sources to payload base
Payload base API now have:

* SetUpSourcesWithTask
* TearDownSourcesWithTask

These are right now implemented only be Live OS payload but when the
sources are ready we can migrate also the Live Image payload.
