+++ pyanaconda/modules/payload/payloads/dnf/dnf.py
@@ -78,3 +78,13 @@ class DNFHandlerModule(PayloadBase):
         """
         # TODO: Implement this method
         pass
+
+    def set_up_sources_with_task(self):
+        """Set up installation sources."""
+        # TODO: Implement this method
+        pass
+
+    def tear_down_sources_with_task(self):
+        """Tear down installation sources."""
+        # TODO: Implement this method
+        pass
